+++ src/Quake/Q_log2.c
@@ -1,8 +1,8 @@
+#include <stdint.h> // int16_t
 
-
-int Q_log2( int x )
+int16_t Q_log2( int16_t x )
 {
-	int answer;
+	int16_t answer;
 
 	answer = 0;
 	while ( (answer >>= 1) != 1 )
