+++ src/Quake/Q_rand.c
@@ -1,6 +1,6 @@
+#include <stdint.h> // int16_t
 
-
-int Q_rand( int *seed )
+int16_t Q_rand( int16_t *seed )
 {
 	*seed = ( 69069 * *seed + 1 );
 	return ( *seed );
