+++ src/Quake/Q_random.c
@@ -1,5 +1,6 @@
+#include <stdint.h> // int16_t
 
-float Q_random( int *seed )
+float Q_random( int16_t *seed )
 {
 	return ( ( Q_rand( seed ) & 0xffff ) / (float)0x10000 );
 }
